requirements.txt
Django==1.7.6
Pillow==2.6.0
PyPDF2==1.24
PyYAML==3.11
WebOb==1.4
WebTest==2.0.17
argh==0.26.1
argparse==1.2.1
beautifulsoup4==4.3.2
cov-core==1.15.0
coverage==3.7.1
diff-match-patch==20121119
dj-database-url==0.3.0
dj-static==0.0.6
django-admin-timeline==1.4
django-admin2==0.6.1
django-adminactions==0.8.5
django-ajax-selects==1.3.5
django-appconf==1.0.1
django-apptemplates==0.0.1
django-autocomplete-light==2.0.9
django-braces==1.4.0
django-classy-tags==0.5.1
django-crispy-forms==1.4.0
django-daterange-filter==1.2.0
django-dynamic-fixture==1.8.1
django-easy-pdf==0.1.0
django-extra-views==0.6.5
django-filter==0.9.2
django-floppyforms==1.3.0
django-geoposition==0.2.2
django-grappelli==2.6.1
django-grappelli-extensions==0.1.3
django-import-export==0.2.6
django-jqmobile==0.1
django-measurement==1.4
django-multifilefield==1.0
django-multiupload==0.3
django-object-tools==1.0.1
django-reporting==0.21
django-toolbelt==0.0.1
django-webtest==1.7.7
django-wkhtmltopdf==2.0.3
djangorestframework==3.1.0
docutils==0.12
evernote==1.25.0
execnet==1.2.0
future==0.15.2
grappelli-dynamic-navbar==0.1.1
gunicorn==19.1.1
html2pdf==0.1
html5lib==0.999
httplib2==0.9.1
measurement==1.7.2
mock==1.0.1
oauth2==1.5.211
pathtools==0.1.2
pdfminer==20140328
pisa==3.0.33
psycopg2==2.5.4
py==1.4.26
pyPdf==1.13
pypdfocr==0.8.2
pytest==2.6.4
pytest-cache==1.0
pytest-cov==1.8.1
pytest-django==2.8.0
pytest-echo==1.3
pytz==2015.6
radar==0.3
reportlab==2.7
selenium==2.44.0
simplejson==3.6.5
six==1.10.0
smartencoding==0.3
static3==0.5.1
sympy==0.7.3
tablib==0.10.0
twentytab-totalsum-admin==0.6
unicodecsv==0.14.1
waitress==0.8.9
watchdog==0.8.3
whitenoise==1.0.6
wsgiref==0.1.2
xhtml2pdf==0.0.6
xlrd==0.9.4
xlwt==0.7.5
xlwt-future==0.8.0
